Change url shortener algorithm. Just randomly generate one
Instead of hashing and encoding it in base32. This would conflict with
the expiration date feature, because a user could try to shorten an url
with an expiration date just to find out that url is already shortened
and the expiration wouldn't take effect.

We could also put the database unique constraint on both the short_url
and expiration_date fields, if we really wanted to hash the url.